made postmatsqrtinv to speed things up slightly
inverse t distribution is slow and will have to be so because of its construction. however part of that was because of calculating the positive definite matrix square root and then inverting it, when the process for making the posmatsqrt make it just as easy to make the inverse. added test for this as well