fix some one dimensional issues in inverted matrix t
because of dropping to 1-D causing leaving matrix stuff (and use of diag()), was having some problems. have an issue which is a problem in general for inverted matrix T - that the "matrixterms" have have a negative determinant for some plausible values. this is not my fault, it's an issue with the distribution itself.