changes matrixt to agree with standard def of density
the matrix variate t is defined inconsistently with the univariate and multivariate t. essentially, those are defined with variance df*dispersion/(df-2) while the matrix variate is defined as dispersion/(df-2). I had written the initial version to agree with the others but am now revising to agree with the standard matrix variate t.