make it so that U and V are more symmetric
had complaints about U and V not being quite symmetric enough
from inv_sympd, turned out to be rounding error on the order
of 1e-13. This is to be expected. So I check if worse than tol,
then, no matter what, make it symmetric. only happens once
per iter so not terribly costly.