fix LDA so it works with BIC/AIC, etc
the issue was that eval.parent() doesn't work when another function
dispatches logLik(), since the parent is then the wrong environment.
I fixed it to just eval() and this worked. Frankly, I worry a little
about this choice, since I don't understand environments well enough
to know the risks, but it works, so, there you go.